+++ src/content-single/EventContentItem/__tests__/EventBanner.tests.js
