+++ src/content-single/EventContentItem/eventIcon.js
